Add Jest and Supertest libraries; Moved server out of app to allow for easier testing.
