+++ config.js
@@ -1,7 +1,9 @@
 module.exports = {
-  addIndex: [{
+
+  createIndex: [{
     indexName: 'reviews_product_id_idx',
     table: 'reviews',
     columns: 'product_id'
   }]
+
 };
