+++ server/__tests__/reviews.test.js
@@ -1,0 +1,10 @@
+const routes = require("../routes");
+
+const request = require("supertest");
+const express = require("express");
+const app = express();
+
+app.use(express.urlencoded({ extended: false }));
+app.use("/", routes);
+
+describe('Reviews ')
