+++ server/admin/createIndex.js
@@ -1,17 +1,26 @@
 const db = require('../db');
-const { createIndex } = require('../../scripts')
+const { createIndex } = require('../../config')
 
-const addIndex = (indexName, table, column) => {
+const createIndexOnTable = (configObj) => {
+  const { indexName, table, column } = configObj;
+
   return db.query(`
     CREATE INDEX $1
     ON $2 ($3);
-  `, [indexName, table, column])
+  `, [ indexName, table, column ])
     .then((res) => {
       console.log(res)
     })
     .catch(err => {
       console.log(err);
     });
+}
+
+
+const addIndexes = (createIndex) => {
+  createIndex.forEach(index => {
+    createIndexOnTable(index)
+  })
 };
 
-addIndex(createIndex)
+addIndexes(createIndex)
